fixing errors P2

+++ src/main/java/co/edu/uniandes/dse/museoartemoderno/services/MuseoService.java
@@ -9,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import co.edu.uniandes.dse.museoartemoderno.repositories.MuseoRepository;
 import co.edu.uniandes.dse.museoartemoderno.entities.MuseoEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.ObraEntity;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.EntityNotFoundException;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.ErrorMessage;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.IllegalOperationException;
@@ -139,6 +140,10 @@ public class MuseoService {
 		
 		if (museoEntity.isEmpty())
 			throw new EntityNotFoundException(ErrorMessage.MUSEO_NOT_FOUND);
+		
+		List<ObraEntity> obras = museoEntity.get().getObras();
+		if (!obras.isEmpty())
+			throw new IllegalOperationException("Unable to delete arista because it has associated obras");
 		
 		museoRepository.deleteById(museoId);
 		log.info("Termina proceso de borrar el museo con id: " + museoId);
